[fix] remove page useless by adding folder component
