+++ frontend/pages/index.tsx
@@ -1,16 +1,18 @@
-import Head from 'next/head';
-import Home from './home';
+import React from 'react';
+import Link from 'next/link';
 
-const Index: React.FC = () => {
+const Home: React.FC = () => {
   return (
-    <div>
-      <Head>
-        <title>Site Hub</title>
-        <link rel="icon" type="image/vnd.icon" href="/image/epitech_logo.ico" />
-      </Head>
-      <Home />
+    <div className="flex flex-col items-center">
+      <h1 className='font-bold'>WELCOME TO THE EPITECH LYON HUB WEBSITE</h1>
+      <div className="flex-box w-4/5">
+        <Link href="/inventory" className='lg:text-3xl md:text-2xl text-md p-4 font-bold'>INVENTORY</Link>
+      </div>
+      <div className="flex-box w-4/5">
+        <Link href="/project" className='lg:text-3xl md:text-2xl text-md p-4 font-bold'>PROJECTS</Link>
+      </div>
     </div>
   );
 };
 
-export default Index;
+export default Home;
